requirements-dev.txt
pytest>=7.0.0
pytest-cov>=4.0.0
pillow>=10.0.0
